* Fix warnings * Document * Move impl
